requirements.txt
Scrapy==2.12.0
psycopg2-binary==2.9.10
redis==5.2.1
pymongo==4.10.1
python-dotenv==1.0.1
pandas==2.2.3
SQLAlchemy==2.0.36